avoid circular dependency between rope op and module
